+++ libraries/helpers.rb
@@ -212,15 +212,10 @@ module MariaDBCookbook
     end
 
     def mariadbbackup_pkg_name
-      case node['platform_family']
-      when 'rhel', 'fedora', 'amazon'
+      if platform_family?('rhel', 'fedora', 'amazon')
         'MariaDB-backup'
       else
-        if new_resource.version == '10.3'
-          'mariadb-backup'
-        else
-          "mariadb-backup-#{new_resource.version}"
-        end
+        new_resource.version == '10.3' ? 'mariadb-backup' : "mariadb-backup-#{new_resource.version}"
       end
     end